libkeepalive: set TCP keepalives options
See:
https://blog.cloudflare.com/when-tcp-sockets-refuse-to-die/
http://codearcana.com/posts/2015/08/28/tcp-keepalive-is-a-lie.html

libkeepalive is a small LD_PRELOAD library to enable TCP keepalives and
TCP_USER_TIMEOUT on any sockets opened by dynamically linked applications,
either outbound (connect(2), using libkeepalive.so) or inbound (listen(2),
using libkeepalive_listen.so).

The typical situation is that a long lasting connection is established
across some network boundary. The connection is idle and some intermediary
device drops the connection state. Enabling TCP keepalives will keep the
connection active and, if the connection is dropped by the intermediary,
force it to be closed.

Alternatively, the connection can be dropped by the intermediary while
there is still pending data. The TCP_USER_TIMEOUT setting will force
the connection to be closed.